[3.0] Move to BIGINT for timestamps
Fixes #8947

+++ Sources/Db/Schema/v3_0/PersonalMessages.php
@@ -73,7 +73,7 @@ class PersonalMessages extends Table
 			),
 			'msgtime' => new Column(
 				name: 'msgtime',
-				type: 'int',
+				type: 'bigint',
 				unsigned: true,
 				not_null: true,
 				default: 0,
